never sleep

+++ lib/subghz/protocols/nice_flor_s.c
@@ -132,9 +132,9 @@ static bool
     FURI_LOG_I(TAG, "counter = %04lX", instance->generic.cnt);
     FURI_LOG_I(TAG, "button = %01lX", btn);
     uint64_t decrypt = ((uint64_t)instance->generic.serial << 16) | instance->generic.cnt;
-    FURI_LOG_I(TAG, "decrypt = %011lX", decrypt);
+    FURI_LOG_I(TAG, "decrypt = %011lX", (uint64_t)decrypt);
     uint64_t enc_part = subghz_protocol_nice_flor_s_encrypt(decrypt, file_name);
-    FURI_LOG_I(TAG, "enc_part = %011lX", enc_part);
+    FURI_LOG_I(TAG, "enc_part = %011lX", (uint64_t)enc_part);
 
     for (int i = 0; i < 16; i++) {
     
@@ -170,7 +170,7 @@ static bool
         //Button 4
         instance->generic.data = ((uint64_t)btn << 4) | (0xF ^ btn ^ loops[3][i]) << 44 | enc_part;
         }
-    FURI_LOG_I(TAG, "key = %013lX", instance->generic.data);
+    FURI_LOG_I(TAG, "key = %013lX", (uint64_t)instance->generic.data);
     
     //Send header
     for(uint8_t i = 35; i > 0; i--) {
@@ -323,7 +323,7 @@ uint64_t subghz_protocol_nice_flor_s_encrypt(uint64_t data, const char* file_nam
     k = ~p[3];
     p[3] = ~p[1];
     p[1] = k;
-    FURI_LOG_I(TAG, "encrypted_data = %011lX", data);
+    FURI_LOG_I(TAG, "encrypted_data = %011lX", (uint64_t)data);
     return data;
 }
 
@@ -362,7 +362,7 @@ static uint64_t
             p[1] = k;
         }
     }
-    FURI_LOG_I(TAG, "decrypted_data = %011lX", data);
+    FURI_LOG_I(TAG, "decrypted_data = %011lX", (uint64_t)data);
     return data;
 }
 
@@ -507,7 +507,7 @@ static void subghz_protocol_nice_flor_s_remote_controller(
         instance->btn = 0;
     } else {
         uint64_t decrypt = subghz_protocol_nice_flor_s_decrypt(instance, file_name);
-        FURI_LOG_I(TAG, "init_decrypted_data = %013lX", decrypt);
+        FURI_LOG_I(TAG, "init_decrypted_data = %013lX", (uint64_t)decrypt);
         instance->cnt = decrypt & 0xFFFF;
         instance->serial = (decrypt >> 16) & 0xFFFFFFF;
         instance->btn = (decrypt >> 48) & 0xF;
